+++ edgy/core/connection/registry.py
@@ -153,6 +153,7 @@ class Registry:
                         name="content_type",
                         owner=model_class,
                         to=real_content_type,
+                        no_constraint=real_content_type.no_constraint,
                     ),
                 )
                 if model_class.meta._is_init: