+++ edgy/core/db/fields/foreign_keys.py
@@ -4,10 +4,14 @@ from typing import TYPE_CHECKING, Any, Callable, Dict, Optional, Sequence, Tuple
 import sqlalchemy
 from pydantic import BaseModel
 
+from edgy.core.db.constants import CASCADE
 from edgy.core.db.fields.base import BaseForeignKey
 from edgy.core.db.fields.factories import ForeignKeyFieldFactory
 from edgy.core.db.fields.types import BaseFieldType
-from edgy.core.db.relationships.relation import OrphanDeletionSingleRelation, SingleRelation
+from edgy.core.db.relationships.relation import (
+    SingleRelation,
+    VirtualCascadeDeletionSingleRelation,
+)
 from edgy.exceptions import FieldDefinitionError
 from edgy.protocols.many_relationship import ManyRelationProtocol
 
@@ -27,7 +31,8 @@ def _removeprefix(text: str, prefix: str) -> str:
 
 
 class BaseForeignKeyField(BaseForeignKey):
-    force_orphan_deletion_relation: bool = False
+    force_cascade_deletion_relation: bool = False
+    relation_has_post_delete_callback: bool = False
 
     def __init__(
         self,
@@ -39,7 +44,7 @@ class BaseForeignKeyField(BaseForeignKey):
         embed_parent: Optional[Tuple[str, str]] = None,
         relation_fn: Optional[Callable[..., ManyRelationProtocol]] = None,
         reverse_path_fn: Optional[Callable[[str], Tuple[Any, str, str]]] = None,
-        delete_orphan: bool = False,
+        remove_referenced: bool = False,
         **kwargs: Any,
     ) -> None:
         self.related_fields = related_fields
@@ -49,13 +54,19 @@ class BaseForeignKeyField(BaseForeignKey):
         self.embed_parent = embed_parent
         self.relation_fn = relation_fn
         self.reverse_path_fn = reverse_path_fn
-        self.delete_orphan = delete_orphan
-        if delete_orphan:
+        self.remove_referenced = remove_referenced
+        if remove_referenced:
             self.post_delete_callback = self._notset_post_delete_callback
         super().__init__(**kwargs)
+        if self.force_cascade_deletion_relation or (
+            self.on_delete == CASCADE and self.no_constraint
+        ):
+            self.relation_has_post_delete_callback = True
 
     async def _notset_post_delete_callback(self, value: Any, instance: "BaseModelType") -> None:
-        await self.expand_relationship(value).delete(delete_orphan_call=True)
+        value = self.expand_relationship(value)
+        if value is not None:
+            await value.delete(remove_referenced_call=True)
 
     async def pre_save_callback(
         self, value: Any, original_value: Any, force_insert: bool, instance: "BaseModelType"
@@ -76,10 +87,10 @@ class BaseForeignKeyField(BaseForeignKey):
     def get_relation(self, **kwargs: Any) -> ManyRelationProtocol:
         if self.relation_fn is not None:
             return self.relation_fn(**kwargs)
-        if self.force_orphan_deletion_relation or (
-            self.on_delete == "CASCADE" and self.no_constraint
+        if self.force_cascade_deletion_relation or (
+            self.on_delete == CASCADE and self.no_constraint
         ):
-            relation: Any = OrphanDeletionSingleRelation
+            relation: Any = VirtualCascadeDeletionSingleRelation
         else:
             relation = SingleRelation
         return cast(
@@ -241,7 +252,7 @@ class BaseForeignKeyField(BaseForeignKey):
         self,
         name: str,
         columns: Sequence[sqlalchemy.Column],
-        schema: Optional[str] = None,
+        schemes: Sequence[str] = (),
         no_constraint: Optional[bool] = None,
     ) -> Sequence[sqlalchemy.Constraint]:
         constraints = []
@@ -249,12 +260,12 @@ class BaseForeignKeyField(BaseForeignKey):
         if not no_constraint:
             target = self.target
             assert not target.__is_proxy_model__
-            prefix = target.meta.tablename
-            if not schema:
-                schema = target.meta.registry.schema.get_default_schema()
-                # if schema and f"{schema}.{prefix}" in target.meta.registry.metadata.tables:
-            if schema:
-                prefix = f"{schema}.{prefix}"
+            # use the last prefix as fallback
+            prefix = ""
+            for schema in schemes:
+                prefix = f"{schema}.{target.meta.tablename}" if schema else target.meta.tablename
+                if prefix in target.meta.registry.metadata.tables:
+                    break
             constraints.append(
                 sqlalchemy.ForeignKeyConstraint(
                     columns,