+++ edgy/core/db/models/base.py
@@ -363,10 +363,10 @@ class EdgyBaseModel(BaseModel, BaseModelType, metaclass=BaseModelMeta):
         assert registry is not None, "registry is not set"
         if metadata is None:
             metadata = registry.metadata
-        if not schema:
-            schema = registry.db_schema
-        if not schema:
-            schema = registry.schema.get_default_schema()
+        schemes: List[str] = []
+        if schema:
+            schemes.append(schema)
+        schemes.append(registry.db_schema or "")
 
         unique_together = cls.meta.unique_together
         index_constraints = cls.meta.indexes
@@ -376,7 +376,7 @@ class EdgyBaseModel(BaseModel, BaseModelType, metaclass=BaseModelMeta):
         for name, field in cls.meta.fields.items():
             current_columns = field.get_columns(name)
             columns.extend(current_columns)
-            global_constraints.extend(field.get_global_constraints(name, current_columns, schema))
+            global_constraints.extend(field.get_global_constraints(name, current_columns, schemes))
 
         # Handle the uniqueness together
         uniques = []
@@ -397,7 +397,7 @@ class EdgyBaseModel(BaseModel, BaseModelType, metaclass=BaseModelMeta):
             *indexes,
             *global_constraints,
             extend_existing=True,
-            schema=schema,
+            schema=schema or registry.db_schema,
         )
 
     @classmethod